fix(relay-electron): exit on uncaught exception, rotate startup log, dedupe IPC contract
Two crash/observability fixes plus the IPC extraction the code asked for:

- uncaughtException handler now flushes the FATAL line and exits(1).
  Registering the handler suppressed Node's default crash-and-terminate,
  so the app kept running in an unknown state -- and blind, since the
  handler had already .end()ed the log stream.
- startup.log rotates the prior launch to startup.log.prev instead of
  truncating on every launch, so 'it crashed yesterday' evidence (the
  whole point of a launch-failure log) survives one relaunch.
- Extract src/shared/ipc-channels.ts as the single source for the 23
  channel names + payload shapes, which were copied verbatim into
  main.ts, preload.ts AND renderer.ts and kept in sync by comment. The
  channel map is a runtime const (main + preload); payload types are
  type-only, so the renderer imports them without runtime coupling.

apps/relay-electron is private; no changeset needed.
